Add backend structure with database models, schemas, and routers for user management, flexibility checks, and statistics. Implement initial database setup and ensure schema creation. Introduce JWT decoding for Firebase authentication and health check endpoint.
